split schema representation JSON codec
